passes tests
Added the Mythical65c24T8 as a third type of 65xx CPU.  It is (so far) the WDC65C02 with two extra opcodes: 0F=CPU to fill A with capabilities and 1F=A24 as a prefix code to switch the next opcode into 24-bit address mode.

Without the prefix code, the CPU acts identically to the WDC65C02 (no Rockwell opcodes).

+++ nmos6502_test.go
@@ -14,7 +14,7 @@ func TestNMOS6502(t *testing.T) {
 }
 
 // To execute test suites from https://github.com/Klaus2m5/6502_65C02_functional_tests
-func executeSuite(t *testing.T, s *State, stepAddress uint16, steps uint8, showStep bool, traceCPUStep uint8) {
+func executeSuite(t *testing.T, s *State, stepAddress uint32, steps uint8, showStep bool, traceCPUStep uint8) {
 	s.reg.setPC(0x0400)
 	currentStep := uint8(255)
 	for {
